test(azure): split responses utils tests
